Start of a 64-bit Windows build.
Does not yet compile, the FnCall interface and the Future might require a newer
version of the compiler, as VS2008 does not have variadic templates or
std::exception_ptr.

+++ Utils/Debug.cpp
@@ -35,7 +35,7 @@ namespace util {
 	static void debugConsoleOutput(const wchar_t *data) {
 		std::string str = String(data).toChar();
 		DWORD result;
-		WriteFile(GetStdHandle(STD_OUTPUT_HANDLE), str.c_str(), str.size(), &result, NULL);
+		WriteFile(GetStdHandle(STD_OUTPUT_HANDLE), str.c_str(), (DWORD)str.size(), &result, NULL);
 	}
 
 	void setDebugTarget(DebugTarget target) {
